add menu component
