+++ components/hero.module.css
@@ -12,61 +12,17 @@
   
   .header{
     position: fixed;
-    top: 30px;
-    left: 30px;
-  }
-  
-  .menu{
-    position: absolute;
     top: 0;
     left: 0;
-    width: 70%;
-    height: 100vh;
-    background-color: rgb(233, 0, 0);
-    transition: all 0.5s ease-in-out;
-    z-index: 99;
-  }
-  
-  .close_btn{
-    margin: 20px;
-    font-size: 50px;
-    font-family: 'Times New Roman', Times, serif;
-  }
-  
-  .menu ul{
-    position: absolute;
-  }
-  
-  .menu li{
-    font-size: 25px;
-    margin: 15px 0;
-    padding: 10px 0;
-    text-decoration: none;
-    border: none;
-    list-style: none;
-  }
-  
-  .menu li:hover{
-    color: white;
-    border-bottom: 1px solid black;
+    padding: 30px;
+    width: 100%;
+    background-color: rgba(0, 0, 0, 0.555);
   }
   
   .logo_wrap{
     position: relative;
     top: 20%;
     text-align: center;
-  }
-  
-  .overlay{
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    background-color: rgba(7, 7, 7, 0.493);
-    backdrop-filter: blur(4px);
-    z-index: 98;
-    transition: all 0.5s ease-in-out;
   }
   
   .top{
@@ -124,56 +80,11 @@
   
     .header{
       position: fixed;
-      top: 30px;
-      left: 30px;
-    }
-  
-    .menu{
-      position: absolute;
-      top: 0;
-      left: 0;
-      width: 400px;
-      height: 100vh;
-      background-color: rgb(233, 0, 0);
-      transition: all 0.5s ease-in-out;
-      z-index: 99;
-    }
-    
-    .close_btn{
-      color: black;
-      margin: 20px;
-      font-size: 60px;
-      font-family: 'Times New Roman', Times, serif;
-    }
-    
-    .menu ul{
-      position: absolute;
-    }
-    
-    .menu li{
-      font-size: 25px;
-      margin: 15px 0;
-      padding: 10px 0;
-      text-decoration: none;
-      border: none;
-      list-style: none;
-    }
-    
-    .menu li:hover{
-      color: white;
-      border-bottom: 1px solid black;
-    }
-  
-    .overlay{
-      position: fixed;
       top: 0;
       left: 0;
       width: 100%;
-      height: 100%;
-      background-color: rgba(7, 7, 7, 0.493);
-      backdrop-filter: blur(4px);
-      z-index: 98;
-      transition: all 0.5s ease-in-out;
+      background-color: rgba(0, 0, 0, 0.555);
+      padding: 30px;
     }
   
     .hamburger_menu{
