+++ components/menu.module.css
@@ -1,1 +1,107 @@
-/* the style for the menu goes here */
+/*================mobile==================*/
+
+@media screen and (max-width: 40rem) {
+
+.menu{
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 70%;
+  height: 100vh;
+  background-color: rgb(233, 0, 0);
+  transition: all 0.5s ease-in-out;
+  z-index: 99;
+}
+
+.close_btn{
+  margin: 20px;
+  font-size: 50px;
+  font-family: 'Times New Roman', Times, serif;
+}
+
+.menu ul{
+  position: absolute;
+}
+
+.menu li{
+  font-size: 25px;
+  margin: 15px 0;
+  padding: 10px 0;
+  text-decoration: none;
+  border: none;
+  list-style: none;
+}
+
+.menu li:hover{
+  color: white;
+  border-bottom: 1px solid black;
+}
+
+.overlay{
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background-color: rgba(7, 7, 7, 0.493);
+  backdrop-filter: blur(4px);
+  z-index: 98;
+  transition: all 0.5s ease-in-out;
+}
+
+}
+
+
+/*================Desk top==================*/
+  
+@media screen and (min-width: 40rem){
+
+  .menu{
+    position: fixed;
+    top: 0;
+    left: 0;
+    width: 400px;
+    height: 100vh;
+    background-color: rgb(233, 0, 0);
+    transition: all 0.5s ease-in-out;
+    z-index: 99;
+  }
+  
+  .close_btn{
+    color: black;
+    margin: 20px;
+    font-size: 60px;
+    font-family: 'Times New Roman', Times, serif;
+  }
+  
+  .menu ul{
+    position: absolute;
+  }
+  
+  .menu li{
+    font-size: 25px;
+    margin: 15px 0;
+    padding: 10px 0;
+    text-decoration: none;
+    border: none;
+    list-style: none;
+  }
+  
+  .menu li:hover{
+    color: white;
+    border-bottom: 1px solid black;
+  }
+
+  .overlay{
+    position: fixed;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    background-color: rgba(7, 7, 7, 0.493);
+    backdrop-filter: blur(4px);
+    z-index: 98;
+    transition: all 0.5s ease-in-out;
+  }
+
+}
